refactor: extract language texts to i18n module
